Add support for GRUB_RECORDFAIL_TIMEOUT option
The GRUB_RECORDFAIL_TIMEOUT option allows users to set a default timeout
value to occur should a system running GRUB2 not receive any input at all at
a prompt. If the timeout value is reached, the system continues to boot
using whatever the first or otherwise selected option is. This is a
particularly useful setting on headless systems where stalling boots can cause
pain.

GRUB_RECORDFAIL_TIMEOUT takes an integer as it's value, which is the time in
seconds to wait before timing out and continuing the boot sequence.

This commit:
  - allows for recordfail_timeout to be set
  - validates that the value of recordfail_timeout is a string
  - adds the config option and value to the GRUB2 config, if set