Only set GRUB_SERIAL_COMMAND if value passed in
Currently, GRUB_SERIAL_COMMAND is passed in with a blank value, and its left
commented out, if no value of GRUB_SERIAL_COMMAND is passed in during either
class instantiation or from Hiera.

There's no need to do this; the key is only relevant when it has a proper
value. It also means that we end up with messy diffs that require some
re-sorting before a diff will make much sense.

Removes the default, a commented key.